pyproject.toml
[project]
name = "web-app"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "coverage>=7.11.0",
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.7",
    "django-crispy-forms>=2.4",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
]
